Updated .c and .h files
Fixed a bug associated with sorting files lexicographically when -a flag is used. Added file printing for multiple directories in lexicographic order. Directories are now printed in lexicographic order. Multiple directory feature is WIP (work in progress). Future plans to change the current printing format and fixing the sorting for the -t flag to sort files based on modification time.